week-1/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
matplotlib>=3.3.0
numpy>=1.21.0
seaborn>=0.11.0
scikit-learn>=0.24.0
fastapi>=0.68.0
uvicorn>=0.15.0
pillow>=8.0.0
jupyter>=1.0.0
requests>=2.25.0